backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
PyJWT==2.8.0
requests==2.31.0
python-dotenv==1.0.0
trycourier==6.2.1
